cessco/requirements.txt
Django==1.5
South==0.8.2
django-model-utils==1.5.0
psycopg2==2.5.1
wsgiref==0.1.2
